Add support for bearer token auth
Adding support to use bearer token as an auth method. Will only use this
method if user enable the config.use_bearer_token configuration.

Added async as we need to make sure the bearer token is available for
the pipelines and pipeline status request. Might be better to pass that
token in instead but will investigate that next time.